Create seed database fxns with faker data + update schema
