+++ package.json
@@ -12,12 +12,14 @@
     "build:prod": "webpack --mode=production",
     "start:dev": "nodemon server/index.js -w server -w database-mysql",
     "test": "jest",
-    "test:coverage": "jest --coverage --colors"
+    "test:coverage": "jest --coverage --colors",
+    "seed": "node database/seed.js"
   },
   "dependencies": {
     "axios": "^0.19.2",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
+    "faker": "^4.1.0",
     "mysql": "^2.18.1",
     "path": "^0.12.7",
     "react": "^16.13.0",
